web/requirements.txt
amqplib==1.0.2
celery==5.4.0
ffmpy==0.5.0
python-postmark==0.6.0
stripe==11.4.1
